ExpectationMaximizationPCA.jl Documentation

ExpectationMaximizationPCA.jl is a Julia rewrite of empca which provides Weighted Expectation Maximization PCA, an iterative method for solving PCA while properly weighting data.

Primary function definitions

The ExpectationMaximizationPCA.EMPCA function is the primary function provided by ExpectationMaximizationPCA.jl

ExpectationMaximizationPCA.EMPCA — Function

EMPCA(μ, n_comp, data, weights; basis_vecs, scores, kwargs...)

Performs expectation-maximization principal component analysis (EMPCA) on data with n_comp basis vectors using weights as the weights. Pre-allocated arrays for basis_vecs, and scores, can be passed via keyword arguments.

ExpectationMaximizationPCA.EMPCA! — Function

EMPCA!(basis_vecs, scores, μ, data_tmp, weights; use_log, kwargs...)

Performs in-place (modifying basis_vecs, scores, and data_tmp) expectation-maximization principal component analysis (EMPCA) on data_tmp using weights as the weights.

Keyword Arguments

use_log::false: whether you want to perform EMPCA on the log of data instead

inds::AbstractUnitRange=axes(basis_vecs, 2): which indices of basis_vecs you want to use

vec_by_vec::Bool=true: whether you want to perform EMPCA one vector at a time (generally preffered) or all at once

Citing EMPCA

If you use ExpectationMaximizationPCA.jl in an academic paper, please include a citation to S. Bailey 2012, PASP, 124, 1015 and optionally an acknowledgement such as:

This work uses the Weighted EMPCA code by Stephen Bailey, available at https://github.com/sbailey/empca/

BibTeX entry:

@ARTICLE{2012PASP..124.1015B,
   author = {{Bailey}, S.},
    title = "{Principal Component Analysis with Noisy and/or Missing Data}",
  journal = {\pasp},
archivePrefix = "arXiv",
   eprint = {1208.4122},
 primaryClass = "astro-ph.IM",
 keywords = {Data Analysis and Techniques},
     year = 2012,
    month = sep,
   volume = 124,
    pages = {1015-1023},
      doi = {10.1086/668105},
   adsurl = {http://adsabs.harvard.edu/abs/2012PASP..124.1015B},
  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
}

Indices

All of the package functions can be found here

Index